requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
frozenlist==1.3.3
googlesearch-python==1.1.0
idna==3.4
multidict==6.0.4
openai==0.27.2
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
soupsieve==2.4
tqdm==4.65.0
twilio==7.17.0
urllib3==1.26.15
utils==1.0.1
yarl==1.8.2
